pyproject.toml
[tool.poetry]
name = "lakers-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "lakers_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
django = "^4.1.6"
django-cors-headers = "^3.14.0"
djoser = "^2.1.0"
django-environ = "^0.10.0"
psycopg2-binary = "^2.9.6"
python-dateutil = "^2.8.2"
mojimoji = "^0.0.12"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
autopep8 = "^2.0.1"
pytest = "^7.3.1"
pytest-django = "^4.5.2"
pytest-mock = "^3.10.0"
pre-commit = "^3.3.3"
isort = "^5.12.0"
black = "^23.7.0"
pylint = "^2.17.4"
pylint-django = "^2.5.3"
django-pandas = "^0.6.6"
pandera = "^0.16.1"
freezegun = "^1.2.2"
geopandas = "^0.13.2"
swifter = "^1.4.0"
pandas = "^2.0.3"
numpy = "^1.25.2"
shapely = "^2.0.1"
pandarallel = "^1.6.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py"]
disable_test_id_escaping_and_forfeit_all_rights_to_community_support = "True"

